implemented galaxy generation

+++ app/models/galaxy.py
@@ -7,7 +7,7 @@ class Galaxy(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(120), nullable=False)
     seed = db.Column(db.Integer, nullable=False)
-    num_stars = db.Column(db.Integer, default=0)
+    num_stars = db.Column(db.BigInteger, default=0)
     type = db.Column(db.String(50))
     meta = db.Column(db.JSON, default={})
     created_at = db.Column(db.DateTime, default=datetime.utcnow)
